chore: refactor app.js to improve structure and add health check route
